refactor(geode-core): group local FE bases under elements + make derham directory-backed
Move the per-element reference bases into a directory-backed `elements`
module and convert `derham` to directory-backed form, mirroring the
deprecated-shim convention established in #378.

- src/p1.rs -> src/elements/p1.rs (elements::p1)
- src/nedelec.rs -> src/elements/nedelec.rs (elements::nedelec)
- src/whitney_face.rs -> src/elements/whitney.rs (elements::whitney,
  pub(crate) preserved, no shim — zero root re-exports)
- src/derham.rs -> src/derham/mod.rs (stays top-level per epic #377
  open-question 4; canonical module path crate::derham unchanged)
- new src/elements/mod.rs with //! group rustdoc

Root flat re-exports for p1/nedelec/derham public items become
#[deprecated] shims pointing at the canonical paths; prelude gains the
high-traffic element + derham items from canonical paths. In-crate lib
callers and intra-doc links migrated to the new paths. geode-core's own
flat-root derham test callers are deferred to cleanup #387.

Pure refactor — no behavior change.

Closes #379
